tayga.c: rework pidfile handling
The whole point of a pidfile, in the absence of a service supervisor, is
for a starting process in a non-interactive job (service) to detect other
instances of that job (service), although not very robustly due to
toctou, pid rollovers. If, at tayga start, the pidfile exists and
contains a PID of an existing process, that is a sign we should bail
out.

Take care to remove the pidfile on exit.
Use the pidfile to determine already-running and concurrently-launched
tayga processes that point to the same pidfile.
Lock the pidfile while writing our PID to it.